Properly utilize $.extend()
Even though the package was extending defaults and options into a new variable (`settings`), settings was never being used.  In some case (where the package was checking both defaults and options), this resulted in merely redundant code: in other cases this could result in buggy code (for example, if the default setting for linked was changed to true at some point in the future, and the user omitted it, it would be treated as false.)

This small PR fixes the branching logic to be a bit more readable and definitely more sustainable.